requirements.txt
alembic==1.11.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
bcrypt==4.1.2
certifi==2024.2.2
click==8.1.7
coverage==7.5.1
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.110.1
fastapi-pagination==0.12.24
flake8==7.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
mccabe==0.7.0
packaging==24.0
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.9
py==1.11.0
pyasn1==0.6.0
pycodestyle==2.11.1
pydantic==2.6.4
pydantic_core==2.16.3
pyflakes==3.2.0
pytest==7.1.3
pytest-lazy-fixture==0.6.3
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqladmin==0.16.1
SQLAlchemy==2.0.29
starlette==0.37.2
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.11.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
WTForms==3.1.2
